Access 'observation' recordings and metadata

query_observation searches for metadata from observation.

query_observation(term = NULL, type = c("sound", "still image"),
cores = 1, pb = TRUE, verbose = TRUE, token = NULL)

Arguments

term

Character vector of length one indicating the species, to query 'observation' database. For example Phaethornis longirostris.

type

Character vector with media type to query for. Currently 'still image' and 'sound' are available.

cores

Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb

Logical argument to control progress bar. Default is TRUE.

verbose

Logical argument that determines if text is shown in console. Default is TRUE.

token

Character refering to the token assigned by Observation.org as authorization for searches.

Value

If all_data is not provided the function returns a data frame with the following media information: id, scientific_name, name, group, group_name, status, rarity, photo, info_text, permalink, determination_requirements, file_url, repository

Details

This function queries for species observation info in the open-access online repository observation. It can return media metadata.

References

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

if (FALSE) {
# search without downloading
View(df1)
}