Subset data frames while preserving attributes

subset_metadata subsets metadata data frames while preserving attributes.

Usage

subset_metadata(metadata, subset, select)

Arguments

metadata

data frame previously obtained from any suwo query function (i.e. `query_reponame()`).

subset

logical expression or numeric vector (with row numbers) indicating rows to keep.

select

logical expression, numeric vector (with column numbers) or character vector (with column names) indicating columns to keep.

Value

returns a data frame which is a subset of the input 'metadata'.

Details

This function subsets a data frame while preserving its attributes. This is important for handling metadata data frames created by query functions in this package, which contain critical attributes that should be retained after subsetting.

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

if (FALSE) { # \dontrun{
# Download metadata
 pant_xc <- query_xenocanto(term = "Phaethornis anthophilus")

# Subset metadata to include only those from Panama
subset_metadata(pant_xc, subset = country == "Panama")

# get first two rows
subset_metadata(pant_xc, subset = 1:2)

# select only specific columns
subset_metadata(pant_xc, select = c("species", "date", "country", "file_url"))
} # }