Access 'taxon codes'

taxon_code_search searches for metadata from clements.

Usage

taxon_code_search(term)

Arguments

term

species name vector of length one indicating the species, to query 'clements database. For example Phaethornis longirostris.

Value

Returns the taxon code for the given species name if match found in the clements taxon code data base

Details

This function queries for taxon code information from the clements taxon code data base clements. It can return taxon code data. Clements data base must be downloaded and available in the work directory.

References

Clements, J. F., P. C. Rasmussen, T. S. Schulenberg, M. J. Iliff, T. A. Fredericks, J. A. Gerbracht, D. Lepage, A. Spencer, S. M. Billerman, B. L. Sullivan, M. Smith, and C. L. Wood. 2024. The eBird/Clements checklist of Birds of the World: v2024.

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

if (FALSE) { # \dontrun{
# search
} # }