Access 'gbif' recordings and metadata

query_gbif searches for metadata from gbif.

Usage

query_gbif(
  term,
  format = c("sound", "image", "video", "interactive resource"),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  dataset = NULL,
  all_data = getOption("all_data", FALSE),
  raw_data = getOption("raw_data", FALSE)
)

Arguments

term

Character vector of length one indicating species, to query the database. For example, Phaethornis longirostris. Required.

format

Character vector with the media format to query for. Options are 'sound', 'image', 'video' and 'interactive resource'. Required.

cores

Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (see options). Note that some repositories might not support parallel queries from the same IP address and errors might occur during the search.

pb

Logical argument to control if progress bar is shown. Default is TRUE. Can be set globally for the current R session via the "pb" option (see options).

verbose

Logical argument that determines if text is shown in console. Default is TRUE. Can be set globally for the current R session via the "verbose" option (see options).

dataset

see https://www.gbif.org/dataset/search?q=.

all_data

Logical argument that determines if all data available from database is shown in the results of search. Default is FALSE. Can be set globally for the current R session via the "all_data" option (see options).

raw_data

Logical argument that determines if the raw data from the API is returned (e.g. without any manipulation). Default is FALSE. Can be set globally for the current R session via the "raw_data" option (see options). If TRUE all_data is set to TRUE internally. Useful for developers, or if users suspect that some data is mishandled during processing (i.e. date information is lost).

Value

If all_data is FALSE the function returns a data frame with the following media information: id, species, date, country, location, latitude, longitude, file_url, repository. If all_data is TRUE the function returns a data frame with the following media information: key, datasetKey, publishingOrgKey, installationKey, hostingOrganizationKey, publishingCountry, protocol, lastCrawled, lastParsed, crawlId, basisOfRecord, occurrenceStatus, taxonKey, kingdomKey, phylumKey, classKey, orderKey, familyKey, genusKey, speciesKey, acceptedTaxonKey, scientificName, acceptedScientificName, kingdom, phylum, order, family, genus, species, genericName, specificEpithet, taxonRank, taxonomicStatus, iucnRedListCategory, dateIdentified, decimalLatitude, decimalLongitude, coordinateUncertaintyInMeters, continent, stateProvince, year, month, day, eventDate, modified, lastInterpreted, references, license, isInCluster, datasetName, recordedBy, identifiedBy, geodeticDatum, class, countryCode, country, rightsHolder, identifier, http...unknown.org.nick, verbatimEventDate, verbatimLocality, collectionCode, gbifID, occurrenceID, taxonID, catalogNumber, institutionCode, eventTime, occurrenceRemarks, http...unknown.org.captive, identificationID, identificationRemarks, distanceFromCentroidInMeters, informationWithheld, sex, lifeStage, preparations, nomenclaturalCode, dynamicProperties, locality, vernacularName, fieldNotes, verbatimElevation, behavior, higherClassification, associatedTaxa, infraspecificEpithet, media-type, media-format, media-references, media-created, media-creator, media-publisher, media-license, media-rightsHolder, file_url, media-description, page, elevation, elevationAccuracy, organismQuantity, organismQuantityType, georeferenceProtocol, verbatimSRS, county, verbatimCoordinateSystem, type, collectionID, individualCount, samplingProtocol, scientificNameID, georeferenceRemarks, language, georeferenceSources, media-title, repository.

Details

This function queries for species observation info in the open-access online repository gbif. It can return media metadata.

References

GBIF.org (2024), GBIF Home Page. Available from: https://www.gbif.org [13 January 2020].

See also

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

if (FALSE) { # \dontrun{
# search sound files
View(df1)
} # }