Access 'Wikiaves' recordings and metadata

query_wikiaves searches for metadata from wikiaves.

Usage

query_wikiaves(
  term,
  format = c("sound", "image"),
  cores = getOption("mc.cores", 1),
  pb = getOption("pb", TRUE),
  verbose = getOption("verbose", TRUE),
  all_data = getOption("all_data", FALSE),
  raw_data = getOption("raw_data", FALSE)
)

Arguments

term

Character vector of length one indicating species, to query the database. For example, Phaethornis longirostris. Required.

format

Character vector with the media format to query for. Options are 'sound' or 'image'. Required.

cores

Numeric vector of length 1. Controls whether parallel computing is applied by specifying the number of cores to be used. Default is 1 (i.e. no parallel computing). Can be set globally for the current R session via the "mc.cores" option (see options). Note that some repositories might not support parallel queries from the same IP address and errors might occur during the search.

pb

Logical argument to control if progress bar is shown. Default is TRUE. Can be set globally for the current R session via the "pb" option (see options).

verbose

Logical argument that determines if text is shown in console. Default is TRUE. Can be set globally for the current R session via the "verbose" option (see options).

all_data

Logical argument that determines if all data available from database is shown in the results of search. Default is FALSE. Can be set globally for the current R session via the "all_data" option (see options).

raw_data

Logical argument that determines if the raw data from the API is returned (e.g. without any manipulation). Default is FALSE. Can be set globally for the current R session via the "raw_data" option (see options). If TRUE all_data is set to TRUE internally. Useful for developers, or if users suspect that some data is mishandled during processing (i.e. date information is lost).

Value

A data frame with the metadata of the observations matching the query. Note that parallel downloading of media files (with the argument cores > 1) does not seem to work on this repository. If all_data is FALSE (default) the data frame contains the following columns: key, species, date, country, locality, latitude, longitude, file_url, repository, format. If all_data is TRUE the data frame contains the following information: recording ID, media type, user ID, species ID, scientific name, common name, repository ID, author, user name, date, verified condition, locality, locality ID, comments, likes, visualizations, url, duration and repository.

Details

This function queries for avian vocalization recordings in the open-access online repository wikiaves. It can return recordings metadata.

References

Schubert, Stephanie Caroline, Lilian Tonelli Manica, and André De Camargo Guaraldo. 2019. Revealing the potential of a huge citizen-science platform to study bird migration. Emu-Austral Ornithology 119.4: 364-373.

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

if (FALSE) { # \dontrun{
# search
df1 <- query_wikiaves(term = "Phaethornis nattereri", format = "image")
View(df1)
} # }