Access 'wikiaves' recordings and metadata

query_wikiaves searches for metadata from wikiaves.

query_wikiaves(term, type, cores = 1, pb = TRUE, verbose = TRUE, all_data = TRUE)

Arguments

term

Character vector of length one indicating species, to query 'wikiaves' database. For example, Phaethornis longirostris.

type

Character vector with media type to query for. Options are 'still image' or 'sound'. Required.

cores

Numeric. Controls whether parallel computing is applied. It specifies the number of cores to be used. Default is 1 (i.e. no parallel computing).

pb

Logical argument to control progress bar. Default is TRUE.

verbose

Logical argument that determines if text is shown in console. Default is TRUE.

all_data

Logical argument that determines if all data available from database is shown in the results of search. Default is TRUE.

Value

If all_data is not provided the function returns a data frame with the following recording information: recording ID, media type, user ID, species ID, scientific name, common name, repository ID, author, user name, date, verified condition, location, location ID, comments, likes, visualizations, url, duration and repository

Details

This function queries for avian vocalization recordings in the open-access online repository wikiaves. It can return recordings metadata.

References

Schubert, Stephanie Caroline, Lilian Tonelli Manica, and André De Camargo Guaraldo. 2019. Revealing the potential of a huge citizen-science platform to study bird migration. Emu-Austral Ornithology 119.4: 364-373.

Author

Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)

Examples

if (FALSE) {
# search
df1 <- query_wikiaves(term = 'Phaethornis nattereri', type = "still image")
View(df1)

}