suwo: A package to streamline bioacoustic analysis

The suwo package aims to simplify the retrieval of nature media (mostly photos, audio files and videos) across multiple online biodiversity databases. The five major media repositories accessed by this package (GBIF, iNaturalist, Macaulay Library, WikiAves, and Xeno-Canto) collectively host more than 250 million media files. Such media are increasingly used in diverse fields, ranging from ecology and evolutionary biology (e.g., trait evolution) to wildlife monitoring and conservation (e.g., for training species detection models). The ability to access and download large amounts of media files and their associated metadata from a single interface thus provides a uniquely powerful resource for facilitating research and conservation efforts.

Details

The main features of the package are:

Obtaining media metadata from online repositories

Downloading associated media files

Updating data sets with new records

License: GPL (>= 2)

See also

Author

Marcelo Araya-Salas, Jorge Elizondo-Calvo & Alejandro Rico-Guevara

Maintainer: Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)