suwo: A package to streamline bioacoustic analysis

suwo is intended to facilitate the search of biological observation data across multiple databases

Details

The main features of the package are:

The use of loops to apply tasks through acoustic signals referenced in a selection table

The production of images in the working folder with spectrograms that allow to organize data and verify acoustic analyzes

The package offers functions to:

Explore and download observational data from GBIF, inaturalist, Xenocanto, Observation.org and Wikiaves

Download different types of data across multiple databases

Consolidate databases and annotation tables

Most of the functions allow the parallelization of tasks, which distributes the tasks among several processors to improve computational efficiency. Tools to evaluate the performance of the analysis at each step are also available. In addition, warbleR satisfies the need for rigorous open source bioacoustic analysis, which facilitates opportunities for use in research and innovation of additional custom analyzes.

License: GPL (>= 2)

See also

Author

Marcelo Araya-Salas & Jorge Elizondo-Calvo

Maintainer: Marcelo Araya-Salas (marcelo.araya@ucr.ac.cr)